refactor: rewrote use[Wrapped|Native]InOrOut hooks
