+++ src/modules/swap/containers/EthFlow/EthFlowBanner.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 
 import { EthFlowBannerContent } from 'modules/swap/pure/EthFlow/EthFlowBanner'
-import { useIsNativeIn } from 'modules/trade/hooks/useIsNative'
+import { useIsNativeIn } from 'modules/trade/hooks/useIsNativeInOrOut'
 import { useWrappedToken } from 'modules/trade/hooks/useWrappedToken'
 
 import useNativeCurrency from 'lib/hooks/useNativeCurrency'
